+++ server/db/config.js
@@ -39,6 +39,8 @@ if (env === 'prod') {
   // use the test database
   config = {
     dbname: "Jarvis_Resources",
+    // dbname: "Jarvis_DEV",
+    // dbname: "Jarvis",
     host: 'SPYGLASS.SRS.IS.KEYSIGHT.COM',
     username: 'JarvisRTUser',
     password: process.env.SPYGLASS_DB_PASSWORD,